+ fixed most warnings for x64 target
git-svn-id: http://miranda.googlecode.com/svn/trunk@9269 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/chat/chat.h
@@ -312,7 +312,7 @@ struct GlobalLogSettings_t {
 	BOOL        TimeStampEventColour;
 	DWORD       dwIconFlags;
 	int         LogTextIndent;
-	int         LoggingLimit;
+	long        LoggingLimit;
 	int         iEventLimit;
 	int         iPopupStyle;
 	int         iPopupTimeout;
